% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fenrir_exacon.R
\name{exacon}
\alias{exacon}
\title{Exact cellwise probabilities of contingency table.}
\usage{
exacon(f)
}
\arguments{
\item{f}{A contingency table.}
}
\value{
A list, with the observed values, expected values, binomial
probabilities, and hypergeometric probabilities.
}
\description{
This function is designed to replicate the behavior of the EXACON program
(Bergman & El-Khouri, 1987, https://doi.org/10.1177/0013164487471024).
}
\examples{
contable <- matrix(c(15, 18, 9, 48), nrow = 2)
exacon(contable)
}
